fix: normalize SysProcAttr and split containment tests
Configure now clears inherited Pgid/Setsid so Start always creates a
new process-group leader. Lifecycle runtime tests move to a separate
file so neither test file exceeds the 300-line subsystem guideline.

Generated-By: looper 0.11.0-beta.1 (runner=fixer, agent=grok-build)
